+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # TBD
 
+# 0.2.1
+### Features
+* Added a new method `WaitForEndpointAvailability` in NetworkContext that can be used to wait until a service's endpoint becomes available 
+
+### Changes
+* Added TypeScript case to regenerate-protobuf-bindings.sh so that now it considers TypeScript within the shell script
+* Generated TypeScript bindings which can now be ouputted by the wrapper script which calls down to developer-tools
+
 # 0.2.0
 ### Changes
 * Replaced the core of `regenerate-protobuf-bindings.sh` with the script from the devtools repo